feat: rename guild to group in prisma schema
